Refactored onboarding_page.dart
